requirements.txt
amqp==5.1.0
async-timeout==4.0.2
autopep8==1.6.0
Babel==2.9.1
billiard==3.6.4.0
blinker==1.4
celery==5.1.1
click
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
Flask==2.0.3
Flask-BabelEx==0.9.4
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.5.1
greenlet
humanize==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.1
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
# pkg_resources==0.0.0
prompt-toolkit==3.0.28
pycodestyle==2.8.0
pyparsing==3.0.7
python-dotenv==0.20.0
pytz==2022.1
redis==4.2.0
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.32
toml==0.10.2
typing_extensions==4.1.1
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.3
wrapt==1.14.0
